Remove pkg/backend package
We no longer use registration pattern for backends, as it is using
globals and init() which we want to avoid using.

We also moved the interface to cli/cmd/cluster package, as it was
the only consumer of it, so actually the entire package can be now
removed, as it no longer has any use.

Refs #992

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>