prometheus_operator: fix labels to ensure endpoints are scraped
Updates the etcd Endpoint and Service label so that etcd metrics
collection is updated to reflect the label change in prometheus-operator

PrometheusRule resource for the components (metallb, rook and contour)
are baed on the label:
```
app: prometheus-operator
```

Once the prometheus-operator is updated, rules referring to the old
labels are not picked up by Prometheus.

This commit updates those resources to the new label:
```
app: kube-prometheus-stack
```

If monitoring is enabled for contour, metallb and rook components, then
those components should also be updated to ensure metrics collection of
the said components is not broken.

Signed-off-by: Imran Pochi <imran@kinvolk.io>