packet: ensure Calico CDRs gets created before use
Currently, bootkube installs all charts in parallel, which causes
calico-host-protection chart installation to fail ocasionally, as it
gets installed before Calico chart itself.

This commit adds a workaround for this, by making an extra copy of
Calico CRDs into bootkube manifests directory, as those will be applied
before Helm charts. This ensures, that the CRDs are created before
calico-host-protection is installed, however it is rather hacky and
should not be considered as final solution for the issue, but a
temporary workaround to ensure cluster bootstrapping remains reliable.

More details in issue #1175.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>